requirements.txt
# RAG framework
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-chroma>=0.1.4
langchain-community>=0.3.0

# Vector DB
chromadb>=0.5.0

# Document parsing
pypdf>=4.3.0
python-docx>=1.1.2

# Search
rank-bm25>=0.2.2
janome>=0.5.0

# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Frontend
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.1
anthropic>=0.40.0
openai>=1.52.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
fpdf2>=2.8.0
